Add rate limit for raider.io

+++ app/schemas/meta_schema.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 class MetaBySpecResponse(BaseModel):
     class_name: str
     spec: str
-    meta: float | None
+    meta: int | None
     spec_type: str
     encounter_id: int
